+++ app/ExternalFiles/Import.php
@@ -521,6 +521,9 @@ class Import extends Directory_Listing_Base {
 			// set date of import (this is not the attachment datetime).
 			$external_file_obj->set_date();
 
+			// set the used service.
+			$external_file_obj->set_service_name( $protocol_handler_obj->get_name() );
+
 			// log that URL has been added as file in media library.
 			$log->create( __( 'URL successfully added in media library.', 'external-files-in-media-library' ), $file_url, 'success', 0, $this->get_identifier() );
 			$log->create( __( 'Using following settings to save this URL in media library:', 'external-files-in-media-library' ) . ' <code>' . wp_json_encode( $file_data ) . '</code>', $file_url, 'info', 2, $this->get_identifier() );