+++ changelog.md
@@ -102,6 +102,7 @@
 - Settings for most services are now saved on user and not global, but can be set to global
 - External sources can now get an individual name
 - ZIP files can not also be opened via any supported TCP protocol
+- Save used service on each external file
 - Optimized URL shortener
 
 ### Fixed